backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pdfminer.six==20231228
langchain==0.2.16
langchain-google-genai==1.0.8
google-generativeai==0.7.2
httpx==0.27.2
python-dotenv==1.0.1
